pyproject.toml
[tool.poetry]
name = "social-media"
version = "0.1.0"
description = ""
authors = ["HulitosCode <heltonmatimbe8@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.17"
sqlalchemy = "^2.0.36"
bcrypt = "^4.2.0"
passlib = "^1.7.4"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
psycopg = "^3.2.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.6"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']


[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']


[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.taskipy.tasks]
run = 'fastapi dev src/main.py'
pre_test = 'task lint'
test = 'pytest --cov=social_media -vv'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
